support SQL-standard date/datetime functions
This allows setting date and datetime type fields to the values `CURRENT_DATE`, `CURRENT_TIMESTAMP`, and `LOCALTIMESTAMP` from the API, and translates that into SQL expressions that get evaluated by the database.

This allows client-side code to set a date or timestamp to current as seen by the server, in case the client computer's clock is incorrect.